chore: create new constants files
