+++ Pharus/Coordinator/Service/Constants/CellReuseIdentifiers.swift
@@ -1,0 +1,14 @@
+//
+//  CellReuseIdentifiers.swift
+//  Pharus
+//
+//  Created by Victor Colen on 15/06/22.
+//
+
+import Foundation
+
+struct CellReuseIdentifiers {
+     let userProjects = "userProjects"
+     let userRankingProjects = "userRankingProjects"
+     let avatarSelection = "avatarSelection"
+}
