+++ Pharus/Coordinator/Service/Constants/DateFormats.swift
@@ -1,0 +1,12 @@
+//
+//  DateFormats.swift
+//  Pharus
+//
+//  Created by Victor Colen on 15/06/22.
+//
+
+import Foundation
+
+struct DateFormats {
+     let projectDateFormat = "dd/MM/yyyy"
+}
